Reorganize transfer helpers for simplicity
Put everything Transfer-related into globus_cli/services/transfer/ and
break up the large module into parts. Attach methods to the custom
client class where appropriate.
