--- src/globus_cli/helpers/__init__.py
@@ -1,5 +1,0 @@
-from .delegate_proxy import fill_delegate_proxy_activation_requirements
-
-__all__ = [
-    "fill_delegate_proxy_activation_requirements",
-]
