+++ src/globus_cli/services/transfer/delegate_proxy.py
