+++ src/globus_cli/services/transfer/endpoint_type.py
